Implement coach customization of grading policy for a CCX
Story #35 As a coach I can see and edit the json for the grading policy.

Story #34 Recalculate grading policy

Repair the broken test for grading by providing an explicit POC context for the request to run in

I am not certain this is the right way to do this, might be better to test the session facility by setting the session for self.client, but this will do for now.