Implement display of CCXs in user-facing templates such as dashboard
Show the display name for the POC in the menu for switching.

Also remove the MOOC: prefix from the MOOC title to avoid truncation issues.

Incorporate the display name of the POC into the display of course org/number/name at the top of the page in the default theme.

This should help to provide a consistent visual representation of which course is being viewed (and if a POC, which one).

provide warning on the grading policy page in system warning style.

refs jazkarta/edx-platform#27

Provide a PocMembership model api that returns PocMemberships for a user.  Default to active memberships.

Build generator that yields triplets of poc, membership and course for each poc membership a user has.

This is an analog of the `get_course_enrollment_pairs` in common/djangoapps/students/views.py and is used in the student dashboard to provide pocs as options for a student to view.

refs jazkarta/edx-platform#26

refactor to use model api instead of inline code.

provide pocs as possible targets for a student to visit from their dashboard.

unset the current active poc when visiting the dashboard so as to avoid needing to switch active pocs when viewing a plain course.  This ensures we can leave template code for the display of courses unchanged.

refs jazkarta/edx-platform#26

update templates to show pocs in a course-like view on a student's dashboard.

revert a portion of the template change so that the poc title will still render properly when a poc is being viewed.