The absolute minimum to get ZeroRPC working with the Enrollment API.
This swaps the api backend of the Enrollment API to point to a ZeroRPC
server. In the real thing, we would make the swapping something that
could happen with configuration... and well, a great many other things.
This is just to prove that it could work.

To run this on devstack:

1. Start the LMS as normal.
    paver lms
2. Start the management command "zserve_enrollment":
    python manage.py lms --settings=devstack zserve_enrollment

Go to the URL: http://localhost:8000/api/enrollment/v1/enrollment