Upgrade to latest selenium.
This will allow us to test against firefox 37, which is an upgrade
over the current browser used for testing (firefox 28). This change will need
to be accompanied by upgrades on different testing environments (e.g., devstack,
jenkins)

The upgrade is allowing us to take advantage of various bug fixes and also
ensure we are using a more relevant browser version.